nlpapp/requirements.txt
flask>=2.0.1,<2.1.0
gunicorn>=20.1.0,<20.2.0
requests
kafka-python
redis
